fix: menu navigation is handled by css
