+++ frontend/package.json
@@ -11,6 +11,7 @@
 	"dependencies": {
 		"class-variance-authority": "^0.7.0",
 		"clsx": "^2.1.1",
+		"lodash": "^4.17.21",
 		"lucide-react": "^0.424.0",
 		"next": "14.2.5",
 		"next-themes": "^0.3.0",
@@ -20,6 +21,7 @@
 		"tailwindcss-animate": "^1.0.7"
 	},
 	"devDependencies": {
+		"@types/lodash": "^4.17.7",
 		"@types/node": "^20",
 		"@types/react": "^18",
 		"@types/react-dom": "^18",
